Changed "Complexity" to "Efficiency", because Big O Notation is a way to measure an algorithm's efficiency.
Big-O classes should not be confused with the actual complexity classes (= P, QP, DLOGTIME, NP, PSPACE,...). Both Big-O and complexity classes are part of Computational Complexity Theory in computer science, thus, "complexity" is half right (but still wrong to some degree). This might seem minor and philosophical, but the terminology can be very confusing for beginners. Please accept this rephrasing in the table headings, it is important to university teaching.